+++ backend/src/test/java/turip/content/service/ContentServiceTest.java
@@ -14,11 +14,14 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.SliceImpl;
 import turip.city.domain.City;
+import turip.content.controller.dto.response.ContentCountResponse;
 import turip.content.controller.dto.response.ContentSearchResponse;
 import turip.content.domain.Content;
 import turip.content.repository.ContentRepository;
 import turip.country.domain.Country;
 import turip.creator.domain.Creator;
+import turip.regioncategory.domain.DomesticRegionCategory;
+import turip.regioncategory.domain.OverseasRegionCategory;
 import turip.tripcourse.service.TripCourseService;
 
 @ExtendWith(MockitoExtension.class)
@@ -62,6 +65,76 @@ class ContentServiceTest {
             // then
             assertThat(contentsByKeyword.loadable())
                     .isFalse();
+        }
+    }
+
+    @DisplayName("지역별 카테고리 기반 컨텐츠 수 조회 기능 테스트")
+    @Nested
+    class CountContentByRegionCategory {
+        @DisplayName("국내 특정 도시(서울) 조회 시 해당 도시의 컨텐츠 수를 반환한다")
+        @Test
+        void countByRegionCategory_withSeoul_returnsSeoulContentCount() {
+            // given
+            String regionCategory = "서울";
+            int expectedCount = 5;
+            given(contentRepository.countByCityName(regionCategory))
+                    .willReturn(expectedCount);
+
+            // when
+            ContentCountResponse response = contentService.countByRegionCategory(regionCategory);
+
+            // then
+            assertThat(response.count()).isEqualTo(expectedCount);
+        }
+
+        @DisplayName("해외 특정 국가(일본) 조회 시 해당 국가의 컨텐츠 수를 반환한다")
+        @Test
+        void countByRegionCategory_withJapan_returnsJapanContentCount() {
+            // given
+            String regionCategory = "일본";
+            int expectedCount = 3;
+            given(contentRepository.countByCityCountryName(regionCategory))
+                    .willReturn(expectedCount);
+
+            // when
+            ContentCountResponse response = contentService.countByRegionCategory(regionCategory);
+
+            // then
+            assertThat(response.count()).isEqualTo(expectedCount);
+        }
+
+        @DisplayName("국내 기타 조회 시 DomesticRegionCategory에 없는 도시들의 컨텐츠 수를 반환한다")
+        @Test
+        void countByRegionCategory_withDomesticEtc_returnsOtherDomesticContentCount() {
+            // given
+            String regionCategory = "국내 기타";
+            int expectedCount = 10;
+            List<String> domesticCategoryNames = DomesticRegionCategory.getDisplayNamesExcludingEtc();
+            given(contentRepository.countByCityNameNotIn(domesticCategoryNames))
+                    .willReturn(expectedCount);
+
+            // when
+            ContentCountResponse response = contentService.countByRegionCategory(regionCategory);
+
+            // then
+            assertThat(response.count()).isEqualTo(expectedCount);
+        }
+
+        @DisplayName("해외 기타 조회 시 OverseasRegionCategory에 없는 국가들의 컨텐츠 수를 반환한다")
+        @Test
+        void countByRegionCategory_withOverseasEtc_returnsOtherOverseasContentCount() {
+            // given
+            String regionCategory = "해외 기타";
+            int expectedCount = 7;
+            List<String> overseasCategoryNames = OverseasRegionCategory.getDisplayNamesExcludingEtc();
+            given(contentRepository.countByCountryNameNotIn(overseasCategoryNames))
+                    .willReturn(expectedCount);
+
+            // when
+            ContentCountResponse response = contentService.countByRegionCategory(regionCategory);
+
+            // then
+            assertThat(response.count()).isEqualTo(expectedCount);
         }
     }
 }
